requirements.txt
fastapi
uvicorn
python-multipart
streamlit
langchain
langchain-community
langchain-huggingface
chromadb
sentence-transformers
beautifulsoup4
python-dotenv
langchain-groq